pyproject.toml
[tool.poetry]
name = "e-learning"
version = "0.1.0"
description = ""
authors = ["VIET-GITL <viet.info.43@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
django = "^5.1.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-cors-headers = "^4.5.0"
pyjwt = "^2.9.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
